Support reconfiguration of number of HARQ processes
In F1, the time until the serving cell config comes from CU can be
longer than when we want to schedule a particular UE. Then, if we don't
configure the default number of HARQ processes, we can have a segfault.

In this commit, we initially configure 8 HARQ processes. When we get the
right number from RRC, we reconfigure.