Revert "kssb < 16 or we cannot do reestablishment: add assert in DU"
This reverts commit eaf28414ca138660a0375d1c0763f2beaeb7a2b1.

This assertion was introduced as we (wrongfully) thought we needed to
calculate the SSB ARFCN at the CU. Since PBCH might carry one bit, which
is not available at the CU, only certain SSB ARFCNs would have been
possible.

With the availability of the SSB ARFCN through the
MeasurementTimingConfiguration, this is not necessary anymore.