Reformulate section on standalone mode
Since tag 2024.w45/commit f54ca5714ed85d1dee2a112fbd3a497e2287cd40,
standalone mode/--sa is the default and does not exist. Correspondingly,
it has been removed from the documentation.

For users reading recent documentation, but running an older version of
OAI, there might be a mismatch in that --sa is necessary to run OAI, but
not given in the documentation. Provide explanatory text to make users
aware of this.