Improve UE SIB1 reception
During cell search, the UE could be at any center frequency for a given
bandwidth to perform a SSB scan. If a MIB is found and the PDCCH0 config
indicates that the SIB1 is outside of current bandwidth, a sync_request
indication is sent to PHY with new carrier config so the UE can sync
with SSB and receive SIB1.

To do this, the function get_type0_PDCCH_CSS_config_parameters() is
disaggregated to smaller functions to fill type0 PDCCH config. After MIB
is decoded, we check if type0 PDCCH config in MIB lies within current
bandwidth in function configure_ue_phy_for_sib1_reception().

Signed-off-by: Sakthivel Velumani <s.velumani@northeastern.edu>