Add error message in sync req handler
Sync request from MAC should either have SSB position that the PHY has to sync
with or instruct the PHY to do a scan. If in SA mode neither is provided, the
UE asserts because likely the MAC had encountered an error and in do-ra/phytest
mode where the RRC config loaded from file could have a configuration that
sends reconfig with sync during initialization and an error message is
displayed.

Signed-off-by: Sakthivel Velumani <s.velumani@northeastern.edu>