fix(test): workaround for channel generation bug in tests
Switches the channel generation in simulation tests from new_channel_desc_scm to a manual method (create_manual_channel_desc). This is a workaround for a possible bug in the new_channel_desc_scm function.

This also includes further bug fixes for the benchmark's batch mode and other logic issues in the scalability tests.