Scale minimum grant PRBs for mixed slot
The config parameter min_grant_prb (default 5) determines the minimum
number of PRBs to be used for UL grant after SR/inactivity, but the TBS
also depends on number of symbols. For a mixed slot, this could mean
that the TBS is overly small.

Avoid this by scaling the number of PRBs for UL grant according to the
symbols, to ensure that even in mixed slot with only 2 symbols, the TBS
is reasonably sized.