TODO: make change to have T_LIB in LOG? Make only T headers available via utils
Ensure that each dependent target of utils gets the logging headers
(depending on T headers), instead of the entire T_LIB. This ensures that
dependent target will have the T headers generated, but ODR violations
because of duplicate symbols due to having T sources linked into utils
cannot happen.

As a consequence, dependencies like directly depending on generate_T are
not necessary for any target that depends on utils.