Workarond for a bug in callback mechanism in XRAN
Update symbol_callback to detect and correct an issue with the calculated
frame index. This is caused by an issue with the timer functionality
in libxran where if a timer is setup close to the GPS second start
the sfn offset `xran_sfn_at_sec_start` is not setup correctly causing
the frame in xran_sense_of_time struct to be returned with an offset
of 100 frames (1 second).