phy: refactor beamforming interface
Use a structure to hold grid resource allocation for each DL and UL slot. It
also holds beam id to be applied for each allocation as passed by MAC via FAPI.
When MAC sends down the allocation info for each PHY channels an UL slot, they
are stored as a list in gNB structure.
1. In case of 7.2 radio and BF is done at the radio, the grid info is sent to
xran after receiving from MAC. This lets xran to send C plane packet for UL
with right info in advance.
2. When BF done in L1 or split 8 case, the stored info is used after receiving
samples to apply beamforming.

For DL slots, the L1 passes the beam and allocation info down to RU after the
samples are produced. And the info is handled based on where the BF is done
like UL slots.

As described above, new file 'phy_digital_beamforming.c' is added to host
function to
1. Store grid info from MAC
2. Process it after scheduler returns
3. Tx and Rx beamforming

Removed beam index from rxdataF and txdataF because the beams are processed
outside of L1 functions where the buffers are used and we must provide only the
logical ports to the L1 functions.

The new beamforming function on Tx path is called using nr_feptx_prec
function pointer in the RU struct for all split radios.