Remove B200-specific clock-rate setting
Avoid this clock-rate setting, as we already do it around
usrp_lib.cpp:1121, so it is superfluous. For instance, we also do not
set a master clock rate for N3xx either.

Also, this has the nice side effect that this enables to synchronize two
B200 using an Octoclock and NTP. Prior to this change, the logs would
print this:

    [HW]   USRP clock set to 1759836386.000000 sec
    [...]
    [HW]   current pps at 3.286750, starting streaming at 4.286750

In other words, the current time setting was reset. With this change
now, we see

    [HW]   USRP clock set to 1759929873.000000 sec
    [HW]   current pps at 1759929873.000000, starting streaming at 1759929874.000000