phy: remove DAS from config
DAS is a special case of digital beamforming. If a logical port must be mapped
to only a subset of antennas in the RU, then this can be achieved by setting
right beam weights. For antennas that should be mapped to the logical port has
unity weights and other antennas are set to 0.

So there is no need to have a separate config in PHY.