build: fix linkage error in physical simulators
nr_dlschsim and nr_ulschsim does not use init_task_ans() which is included in
libthread-pool.a so the symbol is not available in the executables. But
libldpc.so uses init_task_ans() and depends on the main object file to have it.

Last commit adds a .c file to PHY_NR and this might have changed linkage order
and nr_dlschsim and nr_ulschsim no longer have init_task_ans. This commit
explicitly includes adds all symbols from thread-pool to the executable so
libldpc.so can find it.