RRC: Trigger UE Context Setup fallback upon reestablishment on a DU different from the original
As per 38.401 §8.7: if the UE accesses a DU different from the original
one during re-establishment, the CU must trigger UE Context Setup
toward the new DU. This refers to steps 9-10 and steps 9’-10’.
Either one or the other can be executed during re-establishment.

e.g. if this happens during F1 handover, it shall trigger a
UE Context Setup fallback on the CU.

This commit implements the required behavior by:

- Adding a new helper function `rrc_handle_ue_access_different_du()`
  to handle reestablishment on DU different than the original one.
  The function triggers UE Context Setup procedure when the UE
  reestablishes on a different DU.

- Calling this function when detecting that the UE is reestablishing
  on a different DU, i.e.:
  (1) from `handle_rrcReestablishmentComplete()`, as per 38.401 §8.7,
      NOTE on steps 9-10: CU gets UL RRC MESSAGE TRANSFER from a DU
      different than the original
  (2) from `rrc_CU_process_ue_modification_required`, as per 38.401
      §8.7, NOTE on steps 9'-10': CU gets a UE CONTEXT MODIFICATION
      REQUIRED from a DU different from the original one.