Hack for PRACH format B4 O-RU
A hacky solution to make PRACH format B4 work on O-RU. This limits the PRACH
symbol accumulation to first four symbols, making PRACH that much faster. This
is not a proper solution - a proper solution would require to write a symbol
based PRACH receiver. A slight complication is that the PRACH symbol does not
conicide with the OFDM symbl boundary.